requirements.txt
flask
celery
pytube
cloudinary
python-dotenv
redis
pymongo